Move tooltip from ProtocolFormField to InputList
